+++ src/test/groovy/ru/vyarus/dropwizard/orient/https/OrientBinarySslTest.groovy
@@ -1,0 +1,28 @@
+package ru.vyarus.dropwizard.orient.https
+
+import io.dropwizard.testing.junit.DropwizardAppRule
+import org.junit.Rule
+import ru.vyarus.dropwizard.orient.AbstractTest
+import ru.vyarus.dropwizard.orient.support.TestApplication
+import ru.vyarus.dropwizard.orient.support.TestConfiguration
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 27.08.2017
+ */
+class OrientBinarySslTest extends AbstractHttpsTest {
+
+    @Rule
+    DropwizardAppRule<TestConfiguration> RULE =
+            new DropwizardAppRule<TestConfiguration>(TestApplication.class, 'src/test/resources/ru/vyarus/dropwizard/orient/https/sslOrient.yml');
+
+    def "Check orient servlet correctly detect binary ssl"() {
+
+        when: "accessing orient servlet"
+        def data = new URL("http://localhost:8081/orient/").getText()
+        then: "all good"
+        data != null
+        data.contains('<li>Binary port: 2424 (ssl enabled)</li>')
+        data.contains('<li>Http port: 2480</li>')
+    }
+}